[COTLMP] Make "Toggle Mod" a toggleable setting via user config file only
Due to the limitations and technical quirks of COTL UI, the "Toggle Mod" setting is now toggleable via user config file only.
The unloading of the mod at runtime doesn't properly disable the mod as it should, as it created glitches and such in the UI.
Better have this setting be manually changed so that the mod should either be disabled or not during initialization.